Refactoring Runtime.Api - DiagnosticType
Moving out of the `Runtime` class.
