+++ engine/polyglot-api/src/main/scala/org/enso/polyglot/runtime/DiagnosticType.scala
@@ -1,0 +1,13 @@
+package org.enso.polyglot.runtime
+
+import com.github.plokhotnyuk.jsoniter_scala.macros.named
+
+sealed trait DiagnosticType
+
+object DiagnosticType {
+  @named("diagnosticTypeError")
+  case object Error extends DiagnosticType
+
+  @named("diagnosticTypeWarning")
+  case object Warning extends DiagnosticType
+}
